+++ src/legion/ForceFields/mod.rs
@@ -1,3 +1,3 @@
 // SIN!  Sorta INaccurate forcefield.
 
-pub mod ff;
+pub mod SIN;
