+++ src/legion/mod.rs
@@ -1,3 +1,5 @@
-pub mod dynamics;
-pub mod sin;
-pub mod topology;
+pub mod Dynamics;
+pub mod ForceFields;
+pub mod Topology;
+
+use self::ForceFields::SIN;
