+++ src/legion/topology/atom.rs
@@ -3,8 +3,7 @@ use super::particle::HasMass;
 use super::particle::HasPhysics;
 use super::particle::IsSpatial;
 use super::particle::Particle;
-use crate::legion::sin::ff::Elements;
-use crate::legion::sin::ff::ForceField;
+use crate::Legion::ForceFields::SIN::{Elements, ForceField};
 use cgmath::num_traits::ToPrimitive;
 use std::collections::HashMap;
 use std::ops::Sub;
